Update unregister script to track each active service worker found
We aim to remove this component from FT.com so we aim to unregister it from as many of users' browsers as possible. This change adds some tracking code so we can count the number of active service workers in use and know when it has been fully decommissioned. 🐿 v2.12.5