Remove check for cookie length
Allow the service worker to register even if the cookie length is greater than 4000.

Currently many users have cookies longer than 4000 so this will be preventing users from choosing to receive instant alerts as push notifications and in some cases it might be preventing users who have chosen that delivery method from receiving them.

Looking up service worker documentation and examples I can no longer find any reference for the need to check the cookie length. As a lot has changed in browsers and APIs such as service workers in the last 7 years I am hoping that this was a quirk of the API in its early days.