Unregister self
This removes all the service worker code and unregisters itself.

We're doing this because despite trying to unregister the service worker form applications, it's still alive for some browsers (a couple of hundred users).

The internet recommends trying to get the SW to unregister itself and ensure the CDN cache for /__sw.js is purged.